+++ feature/exercise/src/main/java/see/day/exercise/screen/ExerciseDetailScreen.kt
@@ -47,7 +47,7 @@ import see.day.exercise.viewModel.ExerciseDetailViewModel
 import see.day.model.record.RecordType
 import see.day.ui.button.CompleteButton
 import see.day.ui.component.TypeTitle
-import see.day.ui.dialog.DeleteRecordDialog
+import see.day.ui.dialog.ConfirmDialog
 import see.day.ui.dialog.RecordDetailBackDialog
 import see.day.ui.photo.RecordDetailPhotoRow
 import see.day.ui.textField.HealthStat
@@ -124,9 +124,9 @@ fun ExerciseDetailScreenRoot(
 
     var openDeleteDialog by remember { mutableStateOf(false) }
     if (openDeleteDialog) {
-        DeleteRecordDialog(
+        ConfirmDialog(
             onDismiss = { openDeleteDialog = false },
-            onClickDeleteButton = {
+            onClickConfirmButton = {
                 val editMode = uiState.editMode
                 if (editMode is ExerciseDetailUiState.EditMode.Edit) {
                     viewModel.onEvent(ExerciseDetailUiEvent.DeleteRecord(editMode.recordId))
